feat: refactor dtos
