+++ packages/api/src/rooms/dtos/create-room.request.dto.ts
@@ -1,0 +1,6 @@
+import { CreateRoomRequestSchema } from '@/types/rooms/create-room.request.dto';
+import { createZodDto } from '@anatine/zod-nestjs';
+
+export class CreateRoomRequestDto extends createZodDto(
+  CreateRoomRequestSchema
+) {}
