+++ packages/api/src/rooms/dtos/room.response.dto.ts
@@ -1,0 +1,4 @@
+import { RoomResponseSchema } from '@/types/rooms/room.response.dto';
+import { createZodDto } from '@anatine/zod-nestjs';
+
+export class RoomResponseDto extends createZodDto(RoomResponseSchema) {}
