+++ packages/api/src/rooms/rooms.controller.ts
@@ -1,6 +1,6 @@
 import { ClerkAuthGuard } from '@/auth/guards/clerk/clerk.auth.guard';
 import { ApiClerkAuthHeaders } from '@/auth/guards/clerk/open-api-clerk-headers.decorator';
-import { RoomResponseDto } from '@/types/rooms/room.response.dto';
+import { RoomResponseDto } from '@/rooms/dtos/room.response.dto';
 import { Body, Controller, Post, Request, UseGuards } from '@nestjs/common';
 import { ApiCreatedResponse, ApiOperation, ApiTags } from '@nestjs/swagger';
