+++ packages/types/rooms/create-room.request.dto.d.ts
@@ -1,4 +1,4 @@
+import { CreateRoomRequestSchema } from "./create-room.request.dto";
 import { z } from "zod";
-import { CreateRoomRequestSchema } from './create-room.request.dto';
 
-export type CreateRoomRequestDtoType = z.infer<typeof CreateRoomRequestSchema>;
+export type CreateRoomRequestDto = z.infer<typeof CreateRoomRequestSchema>;
