+++ packages/types/rooms/room.response.dto.d.ts
@@ -1,4 +1,4 @@
 import { RoomResponseSchema } from "./room.response.dto";
 import { z } from "zod";
 
-export type RoomResponseDtoType = z.infer<typeof RoomResponseSchema>;
+export type RoomResponseDto = z.infer<typeof RoomResponseSchema>;
